PandocFiltersLiveJuliaCode

Write pandoc filter in Julia

Example

% pandoc --filter=examples/caps.jl -t markdown examples/sample-caps.txt
THIS IS THE CAPS SAMPLE WITH ÄÜÖ.

Library

PandocFiltersLiveJuliaCode.PandocFiltersLiveJuliaCode

PandocFiltersLiveJuliaCode.AST_filter!

PandocFiltersLiveJuliaCode.filter

PandocFiltersLiveJuliaCode.walk!

PandocFiltersLiveJuliaCode.PandocFiltersLiveJuliaCode — Module

PandocFiltersLiveJuliaCode

Package to aid writing Julia scripts that process the pandoc AST serialized as JSON.

PandocFiltersLiveJuliaCode.AST_filter! — Method

AST_filter!(doc, action::Function, format="")
AST_filter!(doc, actions::Vector{Function}, format="")

Converts an action or a list of actions into a filter that reads a JSON-formatted pandoc document from stdin, transforms it by walk!ing the tree with the actions, and returns a new JSON-formatted pandoc document to stdout. The argument is a list of functions action(key, value, format, meta), where key is the type of the pandoc object (e.g. "Str", "Para"), value is the contents of the object (e.g. a string for "Str", a list of inline elements for "Para"), format is the target output format (which will be taken for the first command line argument if present), and meta is the document's metadata. If the function returns None, the object to which it applies will remain unchanged. If it returns an object, the object will be replaced. If it returns a list, the list will be spliced in to the list to which the target object belongs. (So, returning an empty list deletes the object.)

PandocFiltersLiveJuliaCode.filter — Method

filter!(action::Function)
filter!(actions::Vector{Function})

Converts an action or a list of actions into a filter that reads a JSON-formatted pandoc document from stdin, transforms it by walk!ing the tree with the actions, and returns a new JSON-formatted pandoc document to stdout. The argument is a list of functions action(key, value, format, meta), where key is the type of the pandoc object (e.g. "Str", "Para"), value is the contents of the object (e.g. a string for "Str", a list of inline elements for "Para"), format is the target output format (which will be taken for the first command line argument if present), and meta is the document's metadata. If the function returns None, the object to which it applies will remain unchanged. If it returns an object, the object will be replaced. If it returns a list, the list will be spliced in to the list to which the target object belongs. (So, returning an empty list deletes the object.)

PandocFiltersLiveJuliaCode.walk! — Method

walk!(a, action::Function, format, meta)

Function walk! will walk! Pandoc document abstract source tree (AST) and apply filter function on each element of the document AST. Returns a modified tree.

action must be a function which takes four arguments, tag, content, format, meta, and should return

nothing to leave the element unchanged

[] to delete the element

A Pandoc element to replace the element

or a list of Pandoc elements which will be spliced into the list the original object belongs to.